Move profiles to own packages
